Make a struct local
