+++ libgadget/sfr_eff.h
@@ -6,7 +6,6 @@
 #include "timestep.h"
 #include "partmanager.h"
 #include "slotsmanager.h"
-#include "cooling.h"
 
 #define  METAL_YIELD       0.02	/*!< effective metal yield for star formation */
 
@@ -47,25 +46,5 @@ int get_generations(void);
 
 /* Returns 1 if particle is on effective EOS, 0 otherwise*/
 int sfreff_on_eeqos(const struct sph_particle_data * sph, const double a3inv);
-
-/* Structure storing the results of an evaluation of the star formation model*/
-struct sfr_eeqos_data
-{
-    /* Relaxation time*/
-    double trelax;
-    /* Star formation timescale*/
-    double tsfr;
-    /* Internal energy of the gas in the hot phase. */
-    double egyhot;
-    /* Internal energy of the gas in the cold phase.*/
-    double egycold;
-    /* Fraction of the gas in the cold cloud phase. */
-    double cloudfrac;
-    /* Electron fraction after cooling. */
-    double ne;
-};
-
-/* Computes properties of the gas on star forming equation of state*/
-struct sfr_eeqos_data get_sfr_eeqos(struct particle_data * part, struct sph_particle_data * sph, double dtime, const double a3inv, const struct UVBG * const GlobalUVBG);
 
 #endif
